feat: new v3 plugin-updater
