+++ includes/LknwpFilebrowser.php
@@ -1,11 +1,12 @@
 <?php
 
-namespace Lkn\WPFilebrowser;
+namespace Lkn\WPFilebrowser\Includes;
 
-use Lkn\WPFilebrowser\LknwpFilebrowserLoader;
-use Lkn\WPFilebrowser\LknwpFilebrowserI18n;
+use Lkn\WPFilebrowser\Includes\LknwpFilebrowserLoader;
+use Lkn\WPFilebrowser\Includes\LknwpFilebrowserI18n;
 use Lkn\WPFilebrowser\Admin\LknwpFilebrowserAdmin;
 use Lkn\WPFilebrowser\Public\LknwpFilebrowserPublic;
+use Includes\PluginUpdater\Puc\Plugin\Lkn_Puc_Plugin_UpdateChecker;
 
 /**
  * The core plugin class.
@@ -130,9 +131,7 @@ class LknwpFilebrowser {
 
 	public function updater_init()
 	{
-		require_once __DIR__ . '/plugin-updater/plugin-update-checker.php';
-
-		return new \Lkn_Puc_Plugin_UpdateChecker(
+		return new Lkn_Puc_Plugin_UpdateChecker(
 			'https://api.linknacional.com/v3/u/?slug=lknwp-filebrowser',
 			PLUGIN_FILE,
 			'lknwp-filebrowser'
