--- includes/plugin-updater/load-puc.php
@@ -1,6 +1,0 @@
-<?php
-/**
- * Instantiates the autoloader to get the required imports and initial configurations
- */
-require __DIR__ . '/Puc/Autoloader.php';
-new Lkn_Puc_Autoloader();
