--- includes/plugin-updater/plugin-update-checker.php
@@ -1,9 +1,0 @@
-<?php
-/**
- * Plugin Update Checker Library 4.11
- * http://w-shadow.com/
- *
- * Copyright 2021 Janis Elsts
- * Released under the MIT license. See license.txt for details.
- */
-require __DIR__ . '/load-puc.php';
